+++ src/main/java/cn/tangyujun/delines/IDelinesEntity.java
@@ -9,11 +9,11 @@ public interface IDelinesEntity {
 	 * 获取当前实体所在行
 	 * @return 行索引
 	 */
-	int getLineIndex() ;
+	long getLineIndex() ;
 
 	/**
 	 * 设置行索引
 	 * @param lineIndex 行索引
 	 */
-	void setLineIndex(int lineIndex);
+	void setLineIndex(long lineIndex);
 }
